% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the login panel}
\usage{
run_app(
  title = "R Shiny User Management & Authentication",
  pre_login_content = pre_login_content_default,
  url = "https://assets.yanirmor.com",
  footer_message =
    shiny::tags$p("Need help? Contact\\n                                            ",
    shiny::tags$a(href = "mailto:sebsilas@gmail.com", "sebsilas@gmail.com")),
  logged_in_ui = function() {
     shiny::tags$div()
 },
  logged_in_message = "Here is a logged in message!"
)
}
\arguments{
\item{title}{Title of login panel.}

\item{pre_login_content}{Content to go before login.}

\item{url}{Main URL of login panel.}

\item{footer_message}{Message to go in footer.}

\item{logged_in_ui}{UI extra to show on logged in page.}

\item{logged_in_message}{Message to show when someone is logged in.}
}
\description{
Run the login panel
}
