% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the login panel}
\usage{
run_app(
  title = "R Shiny User Management & Authentication",
  pre_login_content = pre_login_content_default,
  url = "https://assets.yanirmor.com",
  footer_message =
    shiny::tags$p("Need help? Contact\\n                                            ",
    shiny::tags$a(href = "mailto:sebsilas@gmail.com", "sebsilas@gmail.com")),
  logged_in_ui = function() {
     shiny::tags$div()
 },
  logged_in_message = "Here is a logged in message!",
  on_start_fun = function() {
     logging::loginfo("Initialising app")
 },
  css_stylesheet = NULL,
  logo_image = "assets/img/icons/user.png",
  icon = "assets/img/icons/user.png",
  on_sign_up_fun = NULL
)
}
\arguments{
\item{title}{Title of login panel.}

\item{pre_login_content}{Content to go before login.}

\item{url}{Main URL of login panel.}

\item{footer_message}{Message to go in footer.}

\item{logged_in_ui}{UI extra to show on logged in page.}

\item{logged_in_message}{Message to show when someone is logged in.}

\item{on_start_fun}{Function to execute on launch}

\item{css_stylesheet}{An optional additional css stylesheet.}

\item{logo_image}{Logo image.}

\item{icon}{Icon.}

\item{on_sign_up_fun}{A function to run when the user signs up. Must have db_con and username arguments.}
}
\description{
Run the login panel
}
